\name{custom.draw.circle}
\alias{custom.draw.circle}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Draws a circle.
}
\description{
Given a center, radius and color, this function draws a circle.
}
\usage{
custom.draw.circle(x, y, r, col)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
X coordinate of the center
}
  \item{y}{
Y coordinate of the center
}
  \item{r}{
Radius of the circle
}
  \item{col}{
Color of the circle
}
}

\value{
This function does not return a value, it just draws a circle.
%% ...
}
\references{
It comes from http://www.r-bloggers.com/circle-packing-with-r/ 
}
\author{
Michael Bedward
}

\examples{
#This is an internal function. No examples required.

## The function is currently defined as
function (x, y, r, col) 
{
    lines(cos(seq(0, 2 * pi, pi/180)) * r + x, sin(seq(0, 2 * 
        pi, pi/180)) * r + y, col = col)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
