% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsetc.R
\name{custom.draw.circle}
\alias{custom.draw.circle}
\title{Draws a circle}
\usage{
custom.draw.circle(x, y, r, col)
}
\arguments{
\item{x}{X coordinate of the center}

\item{y}{Y coordinate of the center}

\item{r}{Radius of the circle}

\item{col}{Color of the circle}
}
\value{
This function does not return a value, it just draws a circle.
}
\description{
Given a center, radius and color, this function draws a circle.
}
\references{
http://www.r-bloggers.com/circle-packing-with-r/
}
\author{
Michael Bedward
}
\keyword{Internal}
