% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microRNA}
\alias{microRNA}
\alias{soilspec}
\alias{Topoisomerase_II_Inhibitors}
\title{Example data sets}
\format{
Each is a dataframe.
}
\usage{
microRNA

soilspec

Topoisomerase_II_Inhibitors
}
\description{
These are data sets used to demonstrate the functions in this package.
}
\examples{
microRNA
soilspec
Topoisomerase_II_Inhibitors
}
\keyword{datasets}
