% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frcc.R
\name{off.diagonal.lambda}
\alias{off.diagonal.lambda}
\title{Calculates the value of the shrinkage coefficient for the off-diagonal matrices}
\usage{
off.diagonal.lambda(xs, p, q)
}
\arguments{
\item{xs}{Matrix with the values for the datasets X and Y.}

\item{p}{Number of variables in the dataset X.}

\item{q}{Number of variables in the dataset Y.}
}
\value{
Shrinkage coefficient for the off-diagonal matrices used to calculate the FRCC canonical structure.
}
\description{
Calculates the value of the shrinkage coefficient for the off-diagonal
matrices as described in [Cruz-Cano et al., 2012]
}
\references{
Cruz-Cano, R.; Lee, M.L.T.; Fast Regularized Canonical Correlation Analysis,
Computational Statistics & Data Analysis, Volume 70, 2014, Pages 88-100,
ISSN 0167-9473, https://doi.org/10.1016/j.csda.2013.09.020.
}
\author{
Raul Cruz-Cano
}
\keyword{Internal}
